Temporary fix for the generics crash. Code is full of debug messages. Parameter parsing still fails.

+++ src/Designite/utils/CSVUtils.java
@@ -11,6 +11,7 @@ public class CSVUtils {
 	//TODO create an integration test for checking the exporting feature
 	public static void initializeCSVDirectory(String projectName) {
 		File dir = new File(Constants.CSV_DIRECTORY_PATH);
+		System.out.println("Initialize dir :: " + dir.getAbsolutePath());
 		createDirIfNotExists(dir);
 		cleanup(dir);
 		initializeNeededFiles(dir);
@@ -27,8 +28,10 @@ public class CSVUtils {
 	}
 	
 	private static void cleanup(File dir) {
-		for (File file : dir.listFiles()) {
-			file.delete();
+		if (dir.listFiles() != null) {
+			for (File file : dir.listFiles()) {
+				file.delete();
+			}
 		}
 	}
 	
@@ -37,6 +40,7 @@ public class CSVUtils {
 		createCSVFile(dir.getPath() + File.separator + Constants.METHOD_METRICS_PATH_SUFFIX, Constants.METHOD_METRICS_HEADER);
 		createCSVFile(dir.getPath() + File.separator + Constants.DESIGN_CODE_SMELLS_PATH_SUFFIX, Constants.DESIGN_CODE_SMELLS_HEADER);
 		createCSVFile(dir.getPath() + File.separator + Constants.IMPLEMENTATION_CODE_SMELLS_PATH_SUFFIX, Constants.IMPLEMENTATION_CODE_SMELLS_HEADER);
+		System.out.println("Created ==> " + dir.listFiles());
 	}
 	
 	private static void createCSVFile(String path, String header) {
